Исправил #4880 Критические сообщения (отказы, накладные)
покрыл тестом новую функциональность.

+++ src/PriceProcessor/Downloader/IMAPHandler.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using LumiSoft.Net.IMAP;
 using LumiSoft.Net.IMAP.Client;
@@ -14,9 +15,24 @@ namespace Inforoom.PriceProcessor.Downloader
 		void PingReader();
 	}
 
+	public class UIDInfo
+	{
+		public virtual int UID { get; private set; }
+
+		public virtual DateTime CreateTime { get; private set; }
+
+		public UIDInfo(int uid)
+		{
+			UID = uid;
+			CreateTime = DateTime.Now;
+		}
+	}
+
 	public class IMAPHandler
 	{
 		public IIMAPReader ImapReader { get; private set; }
+
+		public List<UIDInfo> ErrorInfos { get; set; }
 
 		//Текущее обрабатываемое письмо
 		public Mime Message { get; private set; }
@@ -31,6 +47,7 @@ namespace Inforoom.PriceProcessor.Downloader
 			if (imapReader == null)
 				throw new ArgumentNullException("imapReader");
 			ImapReader = imapReader;
+			ErrorInfos = new List<UIDInfo>();
 		}
 
 		public void ConnectToIMAP()
@@ -54,6 +71,11 @@ namespace Inforoom.PriceProcessor.Downloader
 			return _imapClient.FetchMessages(sequenceMessages, IMAP_FetchItem_Flags.Message, false, true);
 		}
 
+		public bool UIDTimeout(UIDInfo info)
+		{
+			return DateTime.Now.Subtract(info.CreateTime).TotalMinutes > Settings.Default.UIDProcessTimeout;
+		}
+
 		public void ProcessIMAPFolder()
 		{
 			using (var imapClient = new IMAP_Client())
@@ -64,8 +86,11 @@ namespace Inforoom.PriceProcessor.Downloader
 				try
 				{
 					IMAP_FetchItem[] items;
+					var toDelete = new List<IMAP_FetchItem>();
 					do
 					{
+						toDelete.Clear();
+
 						ImapReader.PingReader();
 
 						items = FetchUIDs();
@@ -88,21 +113,29 @@ namespace Inforoom.PriceProcessor.Downloader
 
 								ImapReader.PingReader();
 								ImapReader.ProcessMime(Message);
+								toDelete.Add(item);
 							}
 							catch (Exception ex)
 							{
 								Message = null;
-								ImapReader.ProcessBrokenMessage(item, OneItem, ex);
+								var errorInfo = GetErrorInfo(item.UID);
+								if (UIDTimeout(errorInfo)) {
+									ErrorInfos.Remove(errorInfo);
+									toDelete.Add(item);
+									ImapReader.ProcessBrokenMessage(item, OneItem, ex);
+								}
 							}
 						}
 
 						//Производим удаление писем
-						var sequence = new IMAP_SequenceSet();
+						if (toDelete.Count > 0) {
+							var sequence = new IMAP_SequenceSet();
 
-						sequence.Parse(String.Join(",", items.Select(i => i.UID.ToString()).ToArray()), long.MaxValue);
-						imapClient.DeleteMessages(sequence, true);
+							sequence.Parse(String.Join(",", toDelete.Select(i => i.UID.ToString()).ToArray()), long.MaxValue);
+							imapClient.DeleteMessages(sequence, true);
+						}
 					}
-					while (items != null && items.Length > 0);
+					while (items != null && items.Length > 0 && toDelete.Count > 0);
 				}
 				finally
 				{
@@ -110,6 +143,15 @@ namespace Inforoom.PriceProcessor.Downloader
 				}
 			}
 		}
- 
+
+		private UIDInfo GetErrorInfo(int uid)
+		{
+			var info = ErrorInfos.FirstOrDefault(i => i.UID == uid);
+			if (info == null) {
+				info = new UIDInfo(uid);
+				ErrorInfos.Add(info);
+			}
+			return info;
+		}
 	}
 }
